Pre-inhouse release merge

git-svn-id: https://anonsvn.ncbi.nlm.nih.gov/repos/v1/trunk/c++@39051 78c7ea69-d796-4a43-9a09-de51944f1b03

+++ include/objtools/readers/gene_info/file_utils.hpp
@@ -94,7 +94,7 @@ public:
     static bool CheckExistence(const string& strFile);
 
     /// Get the length of a file, given its name.
-    static int GetLength(const string& strFile);
+    static Int8 GetLength(const string& strFile);
 
     /// Open the given text file for reading.
     static bool OpenTextInputFile(const string& strFileName,
